Collision Resolution Systems now ignore entity order within collision components
At this time, two Collision components are generated when entities a and
b collide - one for the <a, b> pair and another for the <b, a> pair.
This makes sense if, for example, entity a needs to damage b but b also
needs to damage a - these are fundamentally different things, and it
makes no sense not to reuse existing applicability detection
functionality.