Implemented entity existential dependencies
Implemented via a component/system combo, as usual.  The system has been
registered.
NOTE: The graph of entity dependency induced by
DestroyOnParentDestructionComponent links must be a ***TREE***.
NOTE 2: The ordering constraints are unusual for the added system, as it
both marks entities and responds to marks.  This should be handled
properly via recursion.